requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
pydantic==2.9.0
python-dotenv==1.0.1

# Redis for caching and session management
redis==5.0.1

# AI Agent dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Observability and monitoring
langfuse>=2.0.0

# Location processing
geopy>=2.3.0

# Google Maps API integration
googlemaps>=4.7.0

# Text processing and normalization
unidecode>=1.3.0

# Data processing (for vademecum service)
pandas>=1.5.0
pyarrow>=10.0.0

# Dataset management
kagglehub>=0.1.0
